Add ReusableNormalizedNodePruner
This version of NormalizedNodePruner is reusable across invocations,
and improving CPU and memory efficiency in bulk operations due to
not needing to allocate stack nor lookup DataSchemaContextTree.

JIRA: CONTROLLER-1887
(cherry picked from commit 7bef3c9e3f411d3e4abdb5ff43a618537cc9c936)
